SimUtils

Documentation for SimUtils.

SimUtils.Quaternion

SimUtils.gen_pinknoise

SimUtils.mul!

SimUtils.Quaternion — Type

Quaternion{T <: Number}

Lightweight quaternion implementation for simulations.

SimUtils.gen_pinknoise — Function

function gen_pinknoise( beta::Float64=1.0, size::Int64=2^12, dt::Float64=1.0, f0::Float64=1.0 )

description:
> this function generates pink noise with frequency rolloff coefficient 'beta'

positional parameter:
beta             -- type:Float64, frequency rolloff coefficient
size             -- type:Int64, number of points
keyword parameter:
dt               -- type:Float64, the difference of between each data, total time = size * dt, unit: s
f0               -- type:Float64, the spectrum has the shape (f0/f)^(beta), unit: Hz

Note: 
the algorithm is based on 
"Timmer, J. and Koenig, M.: On generating power law noise. Astron. Astrophys. 300, 707-710 (1995)"

SimUtils.mul! — Method

mul!(Q3::Quaternion, Q1::Quaternion{S}, Q2::Quaternion{T}) where {S,T}

Multiplies the quaternions Q1*Q2 and stores the result in-place in Q3. Aliasing of all three arguments is allowed.